news_parser_project/requirements.txt
Django==3.2.6
psycopg2-binary==2.9.1
requests==2.26.0
python-decouple==3.4
django_apscheduler==0.6.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
selenium==4.12.0